+++ frontend/src/components/NotFound.js
@@ -1,5 +1,6 @@
-import React from "react";
-import { Link } from "react-router-dom";
+import React from 'react';
+import { Link } from 'react-router-dom';
+
 const NotFound = () => {
   return (
     <>
