+++ frontend/src/components/styledComponents/Containers.js
@@ -19,6 +19,8 @@ export const ContentContainer = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  border-radius: 1rem;
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 12px;
 
   label {
     display: flex;
